Removed Color+ESMira and moved its code into ContentView and DefaultButtons.
